Refactor: remove auto_route #2
