+++ lib/pages/dashboards/dashboard_page.dart
@@ -1,6 +1,5 @@
 import 'dart:math';
 
-import 'package:auto_route/auto_route.dart';
 import 'package:beamer/beamer.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
@@ -23,7 +22,7 @@ import 'package:lotti/widgets/charts/utils.dart';
 class DashboardPage extends StatefulWidget {
   const DashboardPage({
     super.key,
-    @PathParam() required this.dashboardId,
+    required this.dashboardId,
     this.showBackIcon = true,
   });
