+++ lib/pages/journal/entry_details_page.dart
@@ -1,7 +1,5 @@
 import 'dart:io';
 
-import 'package:auto_route/annotations.dart';
-import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:lotti/classes/journal_entities.dart';
@@ -22,7 +20,7 @@ import 'package:path_provider/path_provider.dart';
 class EntryDetailPage extends StatefulWidget {
   const EntryDetailPage({
     super.key,
-    @PathParam() required this.itemId,
+    required this.itemId,
     this.readOnly = false,
   });
