+++ lib/pages/settings/measurables/measurable_details_page.dart
@@ -1,5 +1,4 @@
 import 'package:adaptive_dialog/adaptive_dialog.dart';
-import 'package:auto_route/auto_route.dart';
 import 'package:enum_to_string/enum_to_string.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
@@ -242,7 +241,7 @@ class _MeasurableDetailsPageState extends State<MeasurableDetailsPage> {
 class EditMeasurablePage extends StatelessWidget {
   EditMeasurablePage({
     super.key,
-    @PathParam() required this.measurableId,
+    required this.measurableId,
   });
 
   final JournalDb _db = getIt<JournalDb>();
