+++ lib/pages/settings/tags/tag_edit_page.dart
@@ -1,4 +1,3 @@
-import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
@@ -258,7 +257,7 @@ class _TagEditPageState extends State<TagEditPage> {
 class EditExistingTagPage extends StatelessWidget {
   EditExistingTagPage({
     super.key,
-    @PathParam() required this.tagEntityId,
+    required this.tagEntityId,
   });
 
   final TagsService tagsService = getIt<TagsService>();
