+++ lib/services/nav_service.dart
@@ -1,6 +1,5 @@
 import 'dart:async';
 
-import 'package:auto_route/auto_route.dart';
 import 'package:flutter/widgets.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/sync/secure_storage.dart';
@@ -13,7 +12,6 @@ class NavService {
   }
 
   String? currentRoute;
-  TabsRouter? tabsRouter;
   List<String> routesByIndex = [];
 
   Future<void> restoreRoute() async {
@@ -25,14 +23,6 @@ class NavService {
         navigateNamedRoute(route);
         debugPrint('restoreRoute: $route');
       });
-    }
-  }
-
-  void bottomNavRouteTap(int index) {
-    final route = routesByIndex[index];
-    debugPrint('bottomNavRouteTap: currentRoute $currentRoute route $route');
-    if ('$currentRoute'.startsWith(route) && route != currentRoute) {
-      tabsRouter?.setActiveIndex(index);
     }
   }
 }
