+++ pubspec.yaml
@@ -1,7 +1,7 @@
 name: lotti
 description: A Smart Journal.
 publish_to: 'none'
-version: 0.8.130+1275
+version: 0.8.130+1276
 
 msix_config:
   display_name: Lotti
@@ -21,7 +21,6 @@ dependencies:
   adaptive_dialog: ^1.3.0
   arb_utils: ^0.2.0
   audio_video_progress_bar: ^0.10.0
-  auto_route: ^4.0.1
   auto_size_text: ^3.0.0
   badges: ^2.0.2
   beamer: ^1.5.2
@@ -160,7 +159,6 @@ dependency_overrides:
 
 dev_dependencies:
   archive: ^3.1.6
-  auto_route_generator: ^4.0.0
   build_runner: ^2.2.0
   dart_code_metrics: ^4.8.1
   drift_dev: ^2.0.2
